adjusted vite config
