+++ .gitignore
@@ -39,4 +39,6 @@ node_modules/
 package-lock.json
 
 frontend/static/js/
+frontend/static/assets/
+
 *.bundle.js
