+++ frontend/templates/base.html
@@ -156,6 +156,6 @@
                 </a>
             </div>
         </footer>
-        <script src="/js/scripts.bundle.js"></script>
+        <script type="module" src="/js/scripts.bundle.js"></script>
     </body>
 </html>
