+++ frontend/templates/loginContent.html
@@ -58,7 +58,7 @@
         </form>
     </div>
 </div>
-<script defer src="/js/login/scripts.bundle.js"></script>
+<script defer type="module" src="/js/login/scripts.bundle.js"></script>
 <div id="register-link" class="mt-5 center" hx-push-url="true" hx-boost="true" hx-target=".content" hx-trigger="click">
     <a class="link-text" href="/register">Register</a>
 </div>
