+++ frontend/templates/registerContent.html
@@ -129,4 +129,4 @@
 <div id="login-link" class="mt-5 center" hx-push-url="true" hx-boost="true" hx-target=".content" hx-trigger="click">
     <a class="link-text" href="/login">Sign in</a>
 </div>
-<script defer src="/js/login/scripts.bundle.js"></script>
+<script defer type="module" src="/js/login/scripts.bundle.js"></script>
